long to int

+++ _print_functions.c
@@ -75,7 +75,7 @@ int print_non_printable(char *s)
  * @n: n.
  * Return: 1.
  */
-int print_number(long int n)
+int print_number(int n)
 {
 	unsigned int x;
 	int printed_characters;
